+++ lib/src/model/notification.dart
@@ -1,3 +1,3 @@
 class WindowsNotification {
-
+  // TODO: WNS Notification Data Model
 }